requirements.txt
absl-py==0.7.1
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-vis==0.4.1
kiwisolver==1.1.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==3.0.5
more-itertools==7.0.0
networkx==2.3
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
pluggy==0.11.0
protobuf==3.7.1
py==1.8.0
pyparsing==2.4.0
pytest==4.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1
requests==2.21.0
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.3
Werkzeug==0.15.2
